EQ2-157 relocate all method in RootController that belong to the Gestionnaire Actor toward GestionnaireRootController
